backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
python-multipart==0.0.9

# LLM 관련
llama-cpp-python==0.2.56

# NumPy 호환성 - 1.x 버전 강제 고정
numpy==1.24.4

# RAG 관련 - 호환성 우선 버전
torch==1.13.1
transformers==4.21.3
sentence-transformers==2.2.2
faiss-cpu==1.7.3

# NLTK 및 NLP 의존성 추가
nltk==3.8.1
click==8.1.7

# InstructorEmbedding 대안 추가
InstructorEmbedding==1.0.1

# 유틸리티
python-dotenv==1.0.1
packaging>=21.0

# 추가 의존성 (안정성 향상)
typing-extensions>=4.8.0
httpx>=0.25.0

# 호환성 보장을 위한 추가 고정
scipy==1.10.1
scikit-learn==1.3.2

# 임베딩 모델 관련 추가 의존성
tokenizers==0.13.3
huggingface-hub==0.16.4
tqdm>=4.64.0
regex>=2022.10.31
requests>=2.28.0
filelock>=3.8.0

# 추가 NLP 라이브러리
joblib>=1.2.0
threadpoolctl>=3.1.0
